# Phase 1: Add comprehensive startup performance tracing (Issue #17)